LibWeb: Split TimeStyleValue out of StyleValue.{h,cpp}
